Added new sql scripts.

+++ analyser/geoip/create_tables.sh
@@ -1,4 +1,12 @@
--- __SCHEMAPLACEHOLDER__ will be replaced by sed for actual schema name
+#!/bin/bash
+
+
+if [[ $# -ne 3 ]]; then
+	echo "$0 <schema name> <user name> <database name>"
+	exit 1;
+fi
+
+sed "s/__SCHEMAPLACEHOLDER__/$1/g" << 'EOF' | psql -U $2 $3 -f- 
 DROP SCHEMA IF EXISTS __SCHEMAPLACEHOLDER__ CASCADE;
 CREATE SCHEMA __SCHEMAPLACEHOLDER__;
 
@@ -24,15 +32,15 @@ $$ LANGUAGE plpgsql;
 
 
 -- MAIN TABLE
-CREATE TABLE domains (
+CREATE TABLE __SCHEMAPLACEHOLDER__.domains (
     id SERIAL PRIMARY KEY,
     fqdn VARCHAR(255) UNIQUE NOT NULL
 );
 
 -- ASN IP TABLE
-CREATE TABLE geoInfo (
+CREATE TABLE __SCHEMAPLACEHOLDER__.geoInfo (
     id SERIAL PRIMARY KEY,
-    host_id INTEGER REFERENCES domains(id),
+    host_id INTEGER REFERENCES __SCHEMAPLACEHOLDER__.domains(id),
     ip VARCHAR(64) NOT NULL,
     city VARCHAR(255) NULL,
     region_name VARCHAR(128) NULL,
@@ -49,3 +57,6 @@ CREATE TABLE geoInfo (
     ipStart VARCHAR(64) NULL,
     ipEnd VARCHAR(64) NULL
 );
+
+EOF
+
